Included model random string save. Also changed some code in optlib_lbfgs (error @ restoration of old gradient information)

+++ InvTlbx_callback_fn/new_iteration.m
@@ -1,5 +1,5 @@
 % function [usr_par] = new_iteration( m, it, usr_par) 
-function [usr_par] = new_iteration( it, m, jm, gm, usr_par) 
+function [usr_par] = new_iteration( it, m, ModRandString, jm, gm, usr_par) 
 % NEW_ITERATION This auxiliary function is called before a new iteration is started.
 % You might use it for writing output files, plotting, ...
 %
@@ -12,9 +12,16 @@ function [usr_par] = new_iteration( it, m, jm, gm, usr_par)
 % Output:
 % usr_par : auxiliary user defined parameters (optional)
 
+
+
 %% initialise
 input_parameters;
-iter = it+1;
+iter = (usr_par.whichFrq-1) * (change_freq_every) + it +1;
+
+disp ' ';
+disp(['---- generating output for iteration ', num2str(iter)]);
+disp ' ';
+
 
 % % from usr_par
 output_path = usr_par.output_path;
@@ -59,6 +66,21 @@ close(fig_mod);
 clearvars('fig_mod');
 
 
+% seismograms
+disp 'plotting seis?'
+% for isrc = 1:nsrc
+%     % determine how many seismograms are actually plotted
+%     if length(vel) > 8; recs = [2:2:length(vel)]; end
+%     
+%     if strcmp(obsPresent, 'yes');
+%         vobs = sEventObs(isrc).vel;
+%         fig_seis(isrc) = plot_seismogram_difference(vel, vobs, t, recs);
+%     else
+%         vobs = make_seismogram_zeros(vel);
+%         fig_seis(isrc) = plot_seismogram_difference(vel, vobs, t, recs, 'nodiff');
+%     end
+% end
+
 % 
 % % gravity kernel (if applicable)
 % if strcmp(use_grav,'yes')
@@ -85,12 +107,14 @@ clearvars('fig_mod');
 % figname = [output_path,'/iter',num2str(iter,'%03d'),'.kernels.abs.rho-vs-vp.png'];
 % print(fig_knl,'-dpng','-r400',figname); close(fig_knl);
 % 
-% % total kernel in (absolute) rhovsvp
-% fig_knl = plot_kernels(K_total, 'rhomulambda',Model, 'total', 'own', 99.95);
-% titel = [project_name,' - iter ',num2str(iter), ' TOTAL kernels (abs rho-mu-lambda)'];
-% mtit(fig_knl,titel, 'xoff', 0.001, 'yoff', 0.04);
-% figname = [output_path,'/iter',num2str(iter,'%03d'),'.kernels-total.abs.rho-mu-lambda.png'];
-% print(fig_knl,'-dpng','-r400',figname); close(fig_knl);
+
+% total kernel in (relative) rhomulambda
+fig_knl = plot_kernels(K_rel, 'rhomulambda', Model, 'total', 'same', 99.95);
+titel = [project_name,' - iter ',num2str(iter), ' TOTAL kernels (rel rho-mu-lambda)'];
+mtit(fig_knl,titel, 'xoff', 0.001, 'yoff', 0.04);
+figname = [output_path,'/iter',num2str(iter,'%03d'),'.kernels-total.rel.rho-mu-lambda.png'];
+print(fig_knl,'-dpng','-r400',figname); close(fig_knl);
+
 % fig_knl = plot_kernels(K_total, 'rhovsvp',Model, 'total', 'own', 99.95);
 % titel = [project_name,' - iter ',num2str(iter), ' TOTAL kernels (abs rho-vs-vp)'];
 % mtit(fig_knl,titel, 'xoff', 0.001, 'yoff', 0.04);
@@ -151,47 +175,63 @@ clearvars('fig_mod');
 %% save output
 
 % if ~exist([output_path,'/iter',num2str(iter,'%03d'),'.all-vars.mat'],'file')
-disp '|  (saving all current variables..)  |'
+disp 'saving all current variables...'
 clearvars('figname', 'savename', 'fig_seisdif', 'fig_mod', ...
     'filenm_old', 'filenm_new', 'fig_knl');
 savename = [output_path,'/iter',num2str(iter,'%03d'),'.all-vars.mat'];
 save(savename, '-regexp', '^(?!(u_fw|v_fw)$).');
 % end
 
-%     
-% %% move all output files to the actual output path
-% blips = dir('./output/*iter.current*');
-% for ii = 1:numel(blips)
-%     bestand = blips(ii).name;
-%     oldfile = ['../output/',bestand];
-%     newfile = [output_path,strrep(blips(ii).name,'iter.current',['iter',num2str(iter,'%03d')])];
-%     movefile(oldfile,newfile);
-% end
-
-%% throw away temp folder
-
-rmdir([output_path,'/fwd_temp'], 's');
-
-%% initialise new frequency if necessary
-
-cfe = change_freq_every;
-whichFrq = floor((iter-1)/cfe)+1;
-if whichFrq > length(sObsPerFreq)
-    whichFrq = length(sObsPerFreq);
+    
+%% move all output files to the actual output path
+blips = dir([output_path,'*iter.current*']);
+for ii = 1:numel(blips)
+    bestand = blips(ii).name;
+    oldfile = [output_path,bestand];
+    newfile = [output_path,strrep(blips(ii).name,'iter.current',['iter',num2str(iter,'%03d')])];
+    movefile(oldfile,newfile);
 end
-
-sEventInfo = usr_par.sObsPerFreq(whichFrq).sEventInfo;
-sEventObs   = usr_par.sObsPerFreq(whichFrq).sEventObs;
 
 
 %% save stuff to usr_par
 
+% iter specific stuff
+usr_par.cumulative_iter = iter;
 usr_par.Model(iter) = Model;
-% usr_par.misfit(iter) = usr_par.misfit_total;
-usr_par.sEventInfo  = sEventInfo;
-usr_par.sEventObs   = sEventObs;
+usr_par.K_rel(iter) = K_rel;
+usr_par.misfit(iter) = jm;
+
+% % frequency specific stuff (used for next iters)
+% usr_par.sEventInfo  = sEventInfo;
+% usr_par.sEventObs   = sEventObs;
 
 
+%% throw away temp folder
 
+TempDir = [output_path,'/fwd_temp'];
+if (exist(TempDir, 'dir'))
+    rmdir([output_path,'/fwd_temp'], 's');
+end
+
+% %% initialise new frequency if necessary
+% 
+% cfe = change_freq_every;
+% whichFrq = floor((iter-1)/cfe)+1;
+% if whichFrq > length(sObsPerFreq)
+%     whichFrq = length(sObsPerFreq);
+% end
+% 
+% sEventInfo = usr_par.sObsPerFreq(whichFrq).sEventInfo;
+% sEventObs   = usr_par.sObsPerFreq(whichFrq).sEventObs;
+
+%% new iteration
+
+
+disp ' ';
+disp '=====================================';
+disp(['==== starting iteration ',num2str(iter+1)]);
+% disp(['==== finished calculations for iter ',num2str(iter)]);
+disp '=====================================';
+disp ' ';
 
 end
